gateway-sgc/src/main/webapp/app/entities/core-sgc/contributor/delete/contributor-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { IContributor } from '../contributor.model';
import { ContributorService } from '../service/contributor.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './contributor-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class ContributorDeleteDialogComponent {
  contributor?: IContributor;

  constructor(protected contributorService: ContributorService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.contributorService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
