gateway-sgc/src/main/webapp/app/entities/core-sgc/contributor/delete/contributor-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ContributorService } from '../service/contributor.service';

import { ContributorDeleteDialogComponent } from './contributor-delete-dialog.component';

describe('Contributor Management Delete Component', () => {
  let comp: ContributorDeleteDialogComponent;
  let fixture: ComponentFixture<ContributorDeleteDialogComponent>;
  let service: ContributorService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, ContributorDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(ContributorDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ContributorDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ContributorService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
